ktl-657 refactor: move layout to blocks
